Refactoring CoreCanvasOptions and CoreTextureOptions to share properties
