+++ packages/fim/src/core/CoreOptions.ts
@@ -5,8 +5,22 @@
 import { FimBitsPerPixel } from '../primitives/FimBitsPerPixel';
 import { FimTextureSampling } from '../primitives/FimTextureSampling';
 
+/** Base interface for `CoreCanvasOptions` and `CoreTextureOptions` */
+export interface CoreOptions {
+  /**
+   * A scaling value ranging from [1.0, 0.0) which causes the underlying canvas to be smaller than the reported
+   * canvas size. A value of 1.0 indicates no downscaling. Downscaling can improve performance and reduce memory
+   * consumption.
+   */
+  readonly downscale: number;
+}
+
+/** Options for `CoreCanvas` and its derived classes */
+export interface CoreCanvasOptions extends CoreOptions {
+}
+
 /** Options for `CoreTexture` */
-export interface CoreTextureOptions {
+export interface CoreTextureOptions extends CoreOptions {
   /** Bits per pixel */
   readonly bpp: FimBitsPerPixel;
 
@@ -18,11 +32,4 @@ export interface CoreTextureOptions {
    * optimizations if the image is not going to be used to store the output of a WebGL shader.
    */
   readonly isReadOnly: boolean;
-
-  /**
-   * A scaling value ranging from [1.0, 0.0) which causes the underlying WebGL texture to be smaller than the reported
-   * texture size. A value of 1.0 indicates no downscaling. Downscaling can improve performance and reduce memory
-   * consumption.
-   */
-  readonly downscale: number;
 }
