+++ packages/fim/src/engine/EngineImage.ts
@@ -13,9 +13,8 @@ import { FimOperation } from '../api/FimOperation';
 import { FimReleaseResourcesFlags } from '../api/FimReleaseResourcesFlags';
 import { FimError, FimErrorCode } from '../primitives/FimError';
 import { CoreCanvas2D } from '../core/CoreCanvas2D';
-import { CoreCanvasOptions } from '../core/CoreCanvasOptions';
+import { CoreCanvasOptions, CoreTextureOptions } from '../core/CoreOptions';
 import { CoreTexture } from '../core/CoreTexture';
-import { CoreTextureOptions } from '../core/CoreTextureOptions';
 import { FimColor } from '../primitives/FimColor';
 import { FimDimensions } from '../primitives/FimDimensions';
 import { FimPoint } from '../primitives/FimPoint';
